pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-chimera"
version = "1.0.0"
description = "Enhanced Master Director - Video Pipeline Control Center"
authors = [{name = "Project Chimera Team"}]
readme = "README_ENHANCED.md"
requires-python = ">=3.8"
dependencies = [
    "Pillow>=9.0.0,<12.0.0",
    "numpy>=1.21.0,<3.0.0",
    "moviepy>=1.0.3,<2.0.0",
    "ollama>=0.1.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "requests>=2.25.0,<3.0.0",
    "schedule>=1.1.0,<2.0.0",
    "pyautogui>=0.9.53,<1.0.0",
    "PyInstaller>=6.10.0,<7.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py38"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "W",   # pycodestyle warnings
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "B006",  # do not use mutable data structures for argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["B011"]

[tool.ruff.lint.isort]
known-first-party = ["core_engine", "content_pipeline", "analysis", "ui"]
known-third-party = ["moviepy", "ollama", "pytrends", "requests", "numpy", "PIL"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "moviepy.*",
    "ollama.*",
    "pytrends.*",
    "gtts.*",
    "piper.*",
    "espeak.*",
    "mido.*",
    "PIL.*",
]
ignore_missing_imports = true
